Added tests for vec<_T, 2> and vec<_T, 4> TODO: decrease the compile times
git-svn-id: file:///usr/local/svn/cpplib/trunk/cftal@969 18305beb-b1e2-0310-bcd8-b14b5419b1d7

+++ x86_v2s64.h
@@ -60,10 +60,10 @@ namespace cftal {
     select(const vec<int64_t, 2>& on_true,
            const vec<int64_t, 2>& on_false);
 
-    template <int64_t _P0, int64_t _P1>
+    template <int _P0, int _P1>
     vec<int64_t, 2> permute(const vec<int64_t, 2>& s);
 
-    template <int64_t _P0, int64_t _P1>
+    template <int _P0, int _P1>
     vec<int64_t, 2> permute(const vec<int64_t, 2>& s0, 
                             const vec<int64_t, 2>& s1);
