+++ x86_v2u64.h
@@ -57,14 +57,14 @@ namespace cftal {
            const vec<uint64_t, 2>& on_false);
 
     template <bool _P0, bool _P1>
-    vec<int64_t, 2> 
+    vec<uint64_t, 2> 
     select(const vec<uint64_t, 2>& on_true,
            const vec<uint64_t, 2>& on_false);
 
-    template <int64_t _P0, int64_t _P1>
+    template <int _P0, int _P1>
     vec<uint64_t, 2> permute(const vec<uint64_t, 2>& s);
 
-    template <int64_t _P0, int64_t _P1>
+    template <int _P0, int _P1>
     vec<uint64_t, 2> permute(const vec<uint64_t, 2>& s0, 
                              const vec<uint64_t, 2>& s1);
